# openapi.yaml
openapi: 3.0.0
info:
  title: AI-Powered Notes System API
  description: |
    A modern RESTful API for note management with AI-powered semantic search and automatic summarization.
    
    **Features:**
    - CRUD operations for notes
    - Semantic search using AI embeddings
    - Automatic note summarization
    - Rate limiting (60 req/minute)
    - Comprehensive input validation
    
    **Authentication:** Not required for MVP
    
    **Base URL:** `http://127.0.0.1:8000/api`
  version: 1.0.0
  contact:
    name: Notes System Team
    url: https://github.com/yourusername/notes-system
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.0.1:8000/api
    description: Local development server
  - url: https://notes-system.example.com/api
    description: Production server
...
paths:
...
  /notes/{id}/summary:
    post:
      tags:
        - AI
      summary: Generate note summary
      description: |
        Generates an AI-powered summary of the note content.
        
        - Mock mode: Extracts first 2 sentences
        - Production mode: Uses GPT-3.5-turbo for quality summaries
      operationId: generateSummary
      parameters:
        - name: id
          in: path
          description: Note ID
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Summary generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  summary:
                    type: string
                    example: "This is a concise summary of the note content. It captures the main ideas in 2-3 sentences."
        '404':
          description: Note not found
        '503':
          description: AI service unavailable
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to generate summary"
                  message:
                    type: string
                    example: "API service error details"
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Optional bearer token for future authentication"
security: []
...